requirements_azure.txt
# Azure Cognitive Search RAG System Requirements
# Core dependencies
azure-search-documents>=11.4.0
azure-core>=1.29.0
azure-identity>=1.15.0

# OpenAI integration
openai>=1.0.0

# Vector operations
numpy>=1.24.0

# Environment management
python-dotenv>=1.0.0

# Optional: For better performance
# azure-storage-blob>=12.19.0  # If using Azure Blob Storage for images
# azure-cognitiveservices-vision-computervision>=0.9.0  # Alternative VLM option
